+++ front-end/src/app/reports/f3x/report-detailed-summary/report-detailed-summary.component.ts
@@ -4,14 +4,14 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { selectReportCodeLabelList } from 'app/store/label-lookup.selectors';
 import { F3xSummary } from 'app/shared/models/f3x-summary.model';
-import { ReportCodeLabelList } from '../../shared/utils/reportCodeLabels.utils';
+import { ReportCodeLabelList } from '../../../shared/utils/reportCodeLabels.utils';
 
 @Component({
-  selector: 'app-report-summary',
-  templateUrl: './report-summary.component.html',
-  styleUrls: ['./report-summary.component.scss'],
+  selector: 'app-report-detailed-summary',
+  templateUrl: './report-detailed-summary.component.html',
+  styleUrls: ['../../style.scss'],
 })
-export class ReportSummaryComponent implements OnInit {
+export class ReportDetailedSummaryComponent implements OnInit {
   report: F3xSummary = new F3xSummary();
   reportCodeLabelList$: Observable<ReportCodeLabelList> = new Observable<ReportCodeLabelList>();
