+++ front-end/src/app/reports/f3x/report-summary/report-summary.component.html
@@ -1,0 +1,105 @@
+<p-card>
+  <div class="summary-header">
+    <h1>Summary page</h1>
+    <div>
+      <strong>FORM 3X:</strong>
+      {{ (this.reportCodeLabelList$ | async | findOnReportCodePipe: report.report_code)?.label }}
+    </div>
+  </div>
+
+  <table class="collapse">
+    <caption>
+      Report Summary
+    </caption>
+    <tr>
+      <th class="right-border">LINE NUMBER</th>
+      <th class="justify-left right-border">LINE DESCRIPTION</th>
+      <th class="right-border">THIS PERIOD</th>
+      <th>CALENDAR YEAR-TO-DATE</th>
+    </tr>
+    <tr>
+      <td class="right-border">6(a)</td>
+      <td class="justify-left right-border">Cash on hand January 1, {{ report.L6a_year_for_above_ytd }}</td>
+      <td class="right-border"></td>
+      <td>{{ report.L6a_cash_on_hand_jan_1_ytd | defaultZero | currency }}</td>
+    </tr>
+    <tr class="gray">
+      <td class="right-border">6(b)</td>
+      <td class="justify-left right-border">Cash on hand at beginning of reporting period</td>
+      <td class="right-border">{{ report.L6b_cash_on_hand_beginning_period | defaultZero | currency }}</td>
+      <td></td>
+    </tr>
+    <tr>
+      <td class="right-border">6(c)</td>
+      <td class="justify-left right-border">Total receipts (from Line 19)</td>
+      <td class="right-border">{{ report.L19_total_receipts_period | defaultZero | currency }}</td>
+      <td>{{ report.L19_total_receipts_ytd | defaultZero | currency }}</td>
+    </tr>
+    <tr class="gray">
+      <td class="right-border">6(d)</td>
+      <td class="justify-left right-border">
+        Subtotal (add Lines 6(b) and 6(c) for Column A and Lines 6(a) and 6(c) from Column B)
+      </td>
+      <td class="right-border">{{ report.L6d_subtotal_period | defaultZero | currency }}</td>
+      <td>{{ report.L6d_subtotal_ytd | defaultZero | currency }}</td>
+    </tr>
+    <tr>
+      <td class="right-border">7</td>
+      <td class="justify-left right-border">Total disbursements (from Line 31)</td>
+      <td class="right-border">{{ report.L7_total_disbursements_period | defaultZero | currency }}</td>
+      <td>{{ report.L7_total_disbursements_ytd | defaultZero | currency }}</td>
+    </tr>
+    <tr class="gray">
+      <td class="right-border">8</td>
+      <td class="justify-left right-border">
+        Cash on hand at close of reporting period (subtract Line 7 from Line 6(d))
+      </td>
+      <td class="right-border">{{ report.L8_cash_on_hand_at_close_period | defaultZero | currency }}</td>
+      <td>{{ report.L8_cash_on_hand_close_ytd | defaultZero | currency }}</td>
+    </tr>
+    <tr>
+      <td class="right-border">9</td>
+      <td class="justify-left right-border">
+        Debts and obligations owed <strong>TO</strong> the committee (Itemize all on Schedule C and/or Schedule D)
+      </td>
+      <td class="right-border">{{ report.L9_debts_to_period | defaultZero | currency }}</td>
+      <td></td>
+    </tr>
+    <tr class="gray last">
+      <td class="right-border">10</td>
+      <td class="justify-left right-border">
+        Debts and obligations owed <strong>BY</strong> the committee (Itemized all on Schedule C and/or Schedule D)
+      </td>
+      <td class="right-border">{{ report.L10_debts_by_period | defaultZero | currency }}</td>
+      <td></td>
+    </tr>
+  </table>
+</p-card>
+
+<div class="grid">
+  <div class="col-4">
+    <div>
+      <button
+        pButton
+        pRipple
+        label="Back"
+        class="p-button-secondary"
+        (click)="router.navigateByUrl('/reports')"
+      ></button>
+    </div>
+    <div class="button-message">PREVIOUS: VALIDATE YOUR REPORT</div>
+  </div>
+  <div class="col-4"></div>
+  <div class="col-4 align-right">
+    <div>
+      <button
+        pButton
+        pRipple
+        label="Review detailed summary page"
+        class="p-button-primary"
+        (click)="router.navigateByUrl('/reports/f3x/detailed-summary/' + report.id)"
+      ></button>
+    </div>
+    <div class="button-message">NEXT: DETAILED SUMMARY PAGE</div>
+  </div>
+</div>
