--- front-end/src/app/reports/report-detailed-summary/report-detailed-summary.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ReportDetailedSummaryComponent } from './report-detailed-summary.component';
-
-describe('ReportDetailedSummaryComponent', () => {
-  let component: ReportDetailedSummaryComponent;
-  let fixture: ComponentFixture<ReportDetailedSummaryComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ ReportDetailedSummaryComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ReportDetailedSummaryComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
