+++ front-end/src/app/reports/style.scss
@@ -6,7 +6,11 @@
   margin-bottom: 0;
 }
 
-.report-summary th {
+table {
+  width: 100%;
+}
+
+th {
   border-top: 2px solid black;
   border-bottom: 3px solid lightgray;
   border-left: 0;
@@ -14,7 +18,7 @@
   padding-right: 10px;
 }
 
-.report-summary td {
+td {
   text-align: center;
   border-top: 0;
   border-bottom: 1px solid lightgray;
@@ -22,20 +26,39 @@
   padding: 10px;
 }
 
-.report-summary .right-border {
+.right-border {
   border-right: 1px solid lightgray;
 }
 
-.report-summary th.justify-left,
-.report-summary td.justify-left {
+th.justify-left,
+td.justify-left {
   text-align: left;
 }
 
-.report-summary .last {
+.last {
   border-bottom: 2px solid black;
 }
 
 .button-message {
   margin-top: 10px;
   color: gray;
+}
+
+caption {
+  font-weight: bolder;
+  text-align: left;
+  margin-bottom: 10px;
+  font-size: 1.2em;
+}
+
+.indent-1 {
+  padding-left: 2em;
+}
+
+.indent-2 {
+  padding-left: 4em;
+}
+
+.table-spacer {
+  height: 3em;
 }
