+++ front-end/src/styles.scss
@@ -60,5 +60,9 @@ label.disabled {
 }
 
 .gray {
-  background-color: #efefef;
+  background-color: #f1f1f1;
+}
+
+.dark-gray {
+  background-color: #aeb0b5;
 }
